counter fix motherfucker

+++ Assets/Scripts/PlayerMovement.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class PlayerMovement : MonoBehaviour
 {
     public float power = 300;
-    private int koks;
+    private float koks;
     public Counter counter;
 	void FixedUpdate()
     {
@@ -40,7 +40,7 @@ public class PlayerMovement : MonoBehaviour
 
         if (sniff) {
             if(Cokeline.IsOverCoke(transform.position.x)){
-                koks += (int)Time.deltaTime;
+                koks += Time.deltaTime;
                 counter.UpdateText(koks);
             }
             else
